backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
google-genai>=1.0.0
python-multipart>=0.0.9
websockets>=13.0.0
aiofiles>=23.2.0
pypdf>=4.0.0
httpx>=0.27.0
motor>=3.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.1
